feat(preflight): move to validate.go and add comments
